move TokenValueView
