+++ src/components/HistoryList/items/ExchangeFromFiatItem.js
@@ -24,7 +24,7 @@ import { useTranslation } from 'translations/translate';
 
 // Components
 import FiatValueView from 'components/display/FiatValueView';
-import TokenValueView from 'components/modern/TokenValueView';
+import TokenValueView from 'components/display/TokenValueView';
 
 // Types
 import type { ExchangeFromFiatEvent } from 'models/History';
