+++ src/components/HistoryList/items/TokenExchangeItem.js
@@ -22,7 +22,7 @@ import * as React from 'react';
 import { useTranslation } from 'translations/translate';
 
 // Components
-import TokenValueView from 'components/modern/TokenValueView';
+import TokenValueView from 'components/display/TokenValueView';
 import { ColumnRight } from 'components/layout/Layout';
 
 // Types
