+++ src/components/display/TokenValueView.js
